Move SpecimenRequestQueryView to study-api
